+++ src/lotus/message/ipc.rs
@@ -1,8 +1,11 @@
 // Copyright 2022-2023 Protocol Labs
 // SPDX-License-Identifier: MIT
 use fvm_shared::clock::ChainEpoch;
+use fvm_shared::econ::TokenAmount;
+use ipc_gateway::Status;
+use ipc_sdk::subnet_id::SubnetID;
 use ipc_subnet_actor::ValidatorSet;
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
 use crate::lotus::message::CIDMap;
 
@@ -28,4 +31,18 @@ pub struct IPCReadGatewayStateResponse {
 pub struct IPCReadSubnetActorStateResponse {
     pub check_period: ChainEpoch,
     pub validator_set: ValidatorSet,
+}
+
+/// SubnetInfo is an auxiliary struct that collects relevant information about the state of a subnet
+#[derive(Debug, Serialize, Deserialize)]
+pub struct SubnetInfo {
+    /// Id of the subnet.
+    pub id: SubnetID,
+    /// Collateral staked in the subnet.
+    #[serde(rename = "stake")]
+    pub collateral: TokenAmount,
+    /// Circulating supply available in the subnet.
+    pub circ_supply: TokenAmount,
+    /// State of the Subnet (Initialized, Active, Killed)
+    pub status: Status,
 }
